pyproject.toml
[project]
name = "meettotext"
version = "0.1.0"
description = ""
authors = [
    {name = "SebastianoTomasi",email = "127763051+SebastianoTomasi@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.3.1,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "imageio-ffmpeg (>=0.6.0,<0.7.0)",
    "openai (>=1.93.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
